+++ conjure-java-core/src/main/java/com/palantir/conjure/java/services/dialogue/DialogueServiceGenerator.java
@@ -19,7 +19,7 @@ package com.palantir.conjure.java.services.dialogue;
 import com.palantir.conjure.java.Options;
 import com.palantir.conjure.java.services.ServiceGenerator;
 import com.palantir.conjure.java.types.TypeMapper;
-import com.palantir.conjure.java.visitor.ClassVisitor;
+import com.palantir.conjure.java.visitor.DialogueClassVisitor;
 import com.palantir.conjure.spec.ConjureDefinition;
 import com.palantir.conjure.spec.TypeDefinition;
 import com.palantir.conjure.visitor.TypeDefinitionVisitor;
@@ -49,14 +49,11 @@ public final class DialogueServiceGenerator implements ServiceGenerator {
 
         TypeMapper parameterTypes = new TypeMapper(
                 conjureDefinition.getTypes(),
-                new DialogueClassNameVisitor(
-                        new ClassVisitor(conjureDefinition.getTypes(), options, ClassVisitor.Mode.PARAMETER),
-                        conjureDefinition.getTypes()));
+                new DialogueClassVisitor(conjureDefinition.getTypes(), options, DialogueClassVisitor.Mode.PARAMETER));
         TypeMapper returnTypes = new TypeMapper(
                 conjureDefinition.getTypes(),
-                new DialogueClassNameVisitor(
-                        new ClassVisitor(conjureDefinition.getTypes(), options, ClassVisitor.Mode.RETURN_VALUE),
-                        conjureDefinition.getTypes()));
+                new DialogueClassVisitor(
+                        conjureDefinition.getTypes(), options, DialogueClassVisitor.Mode.RETURN_VALUE));
         Map<com.palantir.conjure.spec.TypeName, TypeDefinition> typeDefinitionsByName =
                 conjureDefinition.getTypes().stream()
                         .collect(Collectors.toMap(
